+++ mipac/actions/admins/announcement.py
@@ -7,6 +7,7 @@ from mipac.errors.base import ParameterError
 from mipac.http import HTTPClient, Route
 from mipac.models.announcement import Announcement, AnnouncementSystem
 from mipac.types.announcement import IAnnouncement, IAnnouncementSystem
+from mipac.utils.pagination import Pagination
 
 if TYPE_CHECKING:
     from mipac.manager.client import ClientManager
@@ -69,34 +70,31 @@ class AdminAnnouncementActions(AdminAnnouncementClientActions):
         return Announcement(created_announcement, client=self.__client)
 
     async def gets(
-        self, limit: int = 10, since_id: str | None = None, until_id: str | None = None,
+        self,
+        limit: int = 10,
+        since_id: str | None = None,
+        until_id: str | None = None,
+        get_all: bool = False,
     ) -> AsyncGenerator[AnnouncementSystem, None]:
         if limit > 100:
             raise ParameterError('limitは100以下である必要があります')
-        if all:
+        if get_all:
             limit = 100
-
-        async def request(req_body) -> list[AnnouncementSystem]:
-            res: list[IAnnouncementSystem] = await self.__session.request(
-                Route('POST', '/api/admin/announcements/list'), auth=True, json=req_body,
-            )
-            return [AnnouncementSystem(announcement, client=self.__client) for announcement in res]
 
         body = {
             'limit': limit,
             'sinceId': since_id,
             'untilId': until_id,
         }
-        first_req = await request(body)
 
-        for announcement in first_req:
-            yield announcement
-        if all and len(first_req) == 100:
-            body['untilId'] = first_req[-1].id
-            while True:
-                res = await request(body)
-                if len(res) <= 100:
-                    for announcement in res:
-                        yield announcement
-                if len(res) < 100:
-                    break
+        pagination = Pagination[IAnnouncementSystem](
+            self.__session, Route('POST', '/api/admin/announcements/list'), json=body
+        )
+
+        while True:
+            res_annonuncement_systems = await pagination.next()
+            for res_announcement_system in res_annonuncement_systems:
+                yield AnnouncementSystem(res_announcement_system, client=self.__client)
+
+            if get_all is False or pagination.is_final:
+                break