+++ mipac/models/instance.py
@@ -108,3 +108,9 @@ class FederationInstance:
     @property
     def last_communicated_at(self) -> str | None:
         return self.__instance.get('last_communicated_at')
+
+    def __eq__(self, __value: object) -> bool:
+        return isinstance(__value, FederationInstance) and self.id == __value.id
+
+    def __ne__(self, __value: object) -> bool:
+        return not self.__eq__(__value)